[Silver V] Title: 운명, Time: 12 ms, Memory: 2176 KB -BaekjoonHub

+++ 백준/Silver/28282. 운명/운명.cc
@@ -7,7 +7,7 @@ int main(void) {
   cin.tie(0);
   ll X, K;
   cin >> X >> K;
-  ll color;
+  int color;
   ll ans;
   for (int i = 0; i < 2; i++) {
     for (int j = 0; j < X; j++) {
